Fix regression in integer divisions in latest movies category

+++ Emby.Server.Implementations/ScheduledTasks/Tasks/DeleteLogFileTask.cs
@@ -65,7 +65,7 @@ namespace Emby.Server.Implementations.ScheduledTasks.Tasks
 
             foreach (var file in filesToDelete)
             {
-                double percent = index / filesToDelete.Count;
+                double percent = index / (double) filesToDelete.Count;
 
                 progress.Report(100 * percent);
